+++ src/bequiet.spec.ts
@@ -1,7 +1,7 @@
 import {assertEffectsCount, assertSignalsCount} from './assert-helpers.js';
 import {beQuiet} from './bequiet.js';
 import {createSignal, destroySignal} from './createSignal.js';
-import {createEffect} from './effects-api.js';
+import {createEffect} from './effects.js';
 import {touch} from './touch.js';
 
 describe('beQuiet', () => {