+++ src/createSignal.spec.ts
@@ -6,7 +6,7 @@ import {
   muteSignal,
   unmuteSignal,
 } from './createSignal.js';
-import {createEffect} from './effects-api.js';
+import {createEffect} from './effects.js';
 import {touch} from './touch.js';
 
 describe('createSignal', () => {