+++ src/value.ts
@@ -1,5 +1,5 @@
 import {getSignalInstance, isSignal} from './createSignal.js';
-import {queryObjectSignal} from './object-signals-and-effects.js';
+import {findObjectSignalByName} from './object-signals.js';
 import type {SignalLike} from './types.js';
 
 function value<Type>(source: SignalLike<Type>): Type;
@@ -9,7 +9,7 @@ function value<O extends object, K extends keyof O>(source: [O, K]): O[K];
 function value(source: any) {
   return isSignal(source)
     ? (getSignalInstance(source)?.value as any)
-    : (getSignalInstance(queryObjectSignal(...(source as [any, any])))
+    : (getSignalInstance(findObjectSignalByName(...(source as [any, any])))
         ?.value as any);
 }